+++ sdk/appconfiguration/azure-data-appconfiguration/src/test/java/com/azure/data/appconfiguration/ConfigurationClientTestBase.java
@@ -19,6 +19,9 @@ import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.data.appconfiguration.implementation.ConfigurationClientCredentials;
 import com.azure.data.appconfiguration.models.ConfigurationSetting;
+import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
+import com.azure.data.appconfiguration.models.FeatureFlagFilter;
+import com.azure.data.appconfiguration.models.SecretReferenceConfigurationSetting;
 import com.azure.data.appconfiguration.models.SettingFields;
 import com.azure.data.appconfiguration.models.SettingSelector;
 import java.lang.reflect.Field;
@@ -46,6 +49,9 @@ public abstract class ConfigurationClientTestBase extends TestBase {
     private static final int PREFIX_LENGTH = 8;
     private static final int RESOURCE_LENGTH = 16;
     static String connectionString;
+    private static final String FEATURE_FLAG_CONTENT_TYPE = "application/vnd.microsoft.appconfig.ff+json;charset=utf-8";
+    private static final String SECRET_REFERENCE_CONTENT_TYPE =
+        "application/vnd.microsoft.appconfig.keyvaultref+json;charset=utf-8";
 
     private final ClientLogger logger = new ClientLogger(ConfigurationClientTestBase.class);
 
@@ -109,6 +115,48 @@ public abstract class ConfigurationClientTestBase extends TestBase {
     }
 
     @Test
+    public abstract void addFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void addFeatureFlagConfigurationSettingRunner(Consumer<FeatureFlagConfigurationSetting> testRunner) {
+        final String key = getKey();
+
+        List<FeatureFlagFilter> filters = new ArrayList<>();
+
+        filters.add(new FeatureFlagFilter("Microsoft.Percentage")
+                        .setParameters(new HashMap<>() {{
+                            put("Value", "30");
+                        }}));
+
+        final FeatureFlagConfigurationSetting featureSetting =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("Feature Flag X")
+                .setClientFilters(filters)
+                .setContentType(FEATURE_FLAG_CONTENT_TYPE)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+
+        testRunner.accept(featureSetting.setLabel(getLabel()));
+    }
+
+    @Test
+    public abstract void addSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void addSecretReferenceConfigurationSettingRunner(Consumer<SecretReferenceConfigurationSetting> testRunner) {
+        final Map<String, String> tags = new HashMap<>();
+        tags.put("MyTag", "TagValue");
+        tags.put("AnotherTag", "AnotherTagValue");
+
+        final SecretReferenceConfigurationSetting newConfiguration =
+            new SecretReferenceConfigurationSetting(getKey(), "https://localhost")
+                .setKey(getKey())
+                .setTags(tags)
+                .setLabel(getLabel());
+
+        testRunner.accept(newConfiguration);
+    }
+
+    @Test
     public abstract void addConfigurationSettingEmptyKey(HttpClient httpClient,
         ConfigurationServiceVersion serviceVersion);
 
@@ -155,6 +203,67 @@ public abstract class ConfigurationClientTestBase extends TestBase {
 
         testRunner.accept(setConfiguration, updateConfiguration);
         testRunner.accept(setConfiguration.setLabel(label), updateConfiguration.setLabel(label));
+    }
+
+    @Test
+    public abstract void setFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void setFeatureFlagConfigurationSettingRunner(
+        BiConsumer<FeatureFlagConfigurationSetting, FeatureFlagConfigurationSetting> testRunner) {
+        String key = getKey();
+        String label = getLabel();
+
+        final Map<String, String> tags = new HashMap<>();
+        tags.put("MyTag", "TagValue");
+        tags.put("AnotherTag", "AnotherTagValue");
+
+        List<FeatureFlagFilter> filters = new ArrayList<>();
+
+        filters.add(new FeatureFlagFilter("Microsoft.Percentage")
+                        .setParameters(new HashMap<>() {{
+                            put("Value", "30");
+                        }}));
+
+        final FeatureFlagConfigurationSetting featureSetting =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("Feature Flag X")
+                .setClientFilters(filters)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+
+        final FeatureFlagConfigurationSetting updatedFeatureSetting =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("new Feature Flag X")
+                .setClientFilters(filters)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+
+        testRunner.accept(featureSetting.setLabel(label), updatedFeatureSetting.setLabel(label));
+    }
+
+    @Test
+    public abstract void setSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void setSecretReferenceConfigurationSettingRunner(
+        BiConsumer<SecretReferenceConfigurationSetting, SecretReferenceConfigurationSetting> testRunner) {
+        String key = getKey();
+        String label = getLabel();
+
+        final Map<String, String> tags = new HashMap<>();
+        tags.put("MyTag", "TagValue");
+        tags.put("AnotherTag", "AnotherTagValue");
+
+        final SecretReferenceConfigurationSetting newConfiguration =
+            new SecretReferenceConfigurationSetting(getKey(), "https://localhost")
+                .setKey(getKey())
+                .setTags(tags)
+                .setLabel(getLabel());
+        final SecretReferenceConfigurationSetting updatedSetting =
+            new SecretReferenceConfigurationSetting(getKey(), "https://localhost/100")
+                .setKey(getKey())
+                .setTags(tags)
+                .setLabel(getLabel());
+        testRunner.accept(newConfiguration.setLabel(label), updatedSetting.setLabel(label));
     }
 
     @Test
@@ -211,6 +320,39 @@ public abstract class ConfigurationClientTestBase extends TestBase {
     }
 
     @Test
+    public abstract void getFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void getFeatureFlagConfigurationSettingRunner(Consumer<FeatureFlagConfigurationSetting> testRunner) {
+        String key = getKey();
+        List<FeatureFlagFilter> filters = new ArrayList<>();
+        filters.add(new FeatureFlagFilter("Microsoft.Percentage")
+                        .setParameters(new HashMap<>() {{
+                            put("Value", "30");
+                        }}));
+
+        final FeatureFlagConfigurationSetting featureSetting =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("Feature Flag X")
+                .setClientFilters(filters)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+
+        testRunner.accept(featureSetting.setLabel("myLabel"));
+    }
+
+    @Test
+    public abstract void getSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void getSecretReferenceConfigurationSettingRunner(Consumer<SecretReferenceConfigurationSetting> testRunner) {
+        final SecretReferenceConfigurationSetting newConfiguration =
+            new SecretReferenceConfigurationSetting(getKey(), "https://localhost")
+                .setKey(getKey())
+                .setLabel(getLabel());
+        testRunner.accept(newConfiguration);
+    }
+
+    @Test
     public abstract void getConfigurationSettingNotFound(HttpClient httpClient,
         ConfigurationServiceVersion serviceVersion);
 
@@ -229,6 +371,44 @@ public abstract class ConfigurationClientTestBase extends TestBase {
 
         testRunner.accept(deletableConfiguration);
         testRunner.accept(deletableConfiguration.setLabel(label));
+    }
+
+    @Test
+    public abstract void deleteFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void deleteFeatureFlagConfigurationSettingRunner(Consumer<FeatureFlagConfigurationSetting> testRunner) {
+        String key = getKey();
+        String label = getLabel();
+
+        List<FeatureFlagFilter> filters = new ArrayList<>();
+        filters.add(new FeatureFlagFilter("Microsoft.Percentage")
+                        .setParameters(new HashMap<>() {{
+                            put("Value", "30");
+                        }}));
+
+        final FeatureFlagConfigurationSetting featureSetting =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("Feature Flag X")
+                .setClientFilters(filters)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+
+        testRunner.accept(featureSetting.setLabel(label));
+    }
+
+    @Test
+    public abstract void deleteSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void deleteSecretReferenceConfigurationSettingRunner(Consumer<SecretReferenceConfigurationSetting> testRunner) {
+        String key = getKey();
+        String label = getLabel();
+
+        final SecretReferenceConfigurationSetting newConfiguration =
+            new SecretReferenceConfigurationSetting(key, "https://localhost")
+                .setKey(getKey())
+                .setLabel(getLabel());
+        testRunner.accept(newConfiguration.setLabel(label));
     }
 
     @Test
@@ -281,6 +461,50 @@ public abstract class ConfigurationClientTestBase extends TestBase {
 
         final ConfigurationSetting lockConfiguration = new ConfigurationSetting().setKey(key).setValue("myValue");
         testRunner.accept(lockConfiguration);
+    }
+
+    @Test
+    public abstract void setReadOnlyWithFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    @Test
+    public abstract void clearReadOnlyWithFeatureFlagConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void lockUnlockFeatureFlagRunner(Consumer<FeatureFlagConfigurationSetting> testRunner) {
+        String key = getKey();
+
+        List<FeatureFlagFilter> filters = new ArrayList<>();
+        filters.add(new FeatureFlagFilter("Microsoft.Percentage")
+                        .setParameters(new HashMap<>() {{
+                            put("Value", "30");
+                        }}));
+
+        final FeatureFlagConfigurationSetting lockConfiguration =
+            new FeatureFlagConfigurationSetting(key, false)
+                .setDisplayName("Feature Flag X")
+                .setClientFilters(filters)
+                .setValue(getFeatureFlagConfigurationSettingValue(key));
+        testRunner.accept(lockConfiguration);
+    }
+
+    @Test
+    public abstract void setReadOnlyWithSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    @Test
+    public abstract void clearReadOnlyWithSecretReferenceConfigurationSettingConvenience(HttpClient httpClient,
+        ConfigurationServiceVersion serviceVersion);
+
+    void lockUnlockSecretReferenceRunner(Consumer<SecretReferenceConfigurationSetting> testRunner) {
+        String key = getKey();
+
+        final SecretReferenceConfigurationSetting newConfiguration =
+            new SecretReferenceConfigurationSetting(key, "https://localhost")
+                .setKey(getKey())
+                .setLabel(getLabel());
+
+        testRunner.accept(newConfiguration);
     }
 
     @Test
@@ -480,6 +704,25 @@ public abstract class ConfigurationClientTestBase extends TestBase {
         assertConfigurationEquals(expected, response, 200);
     }
 
+    static void assertFeatureFlagConfigurationSettingEquals(FeatureFlagConfigurationSetting expected,
+        FeatureFlagConfigurationSetting actual) {
+        assertEquals(expected.getFeatureId(), actual.getFeatureId());
+        assertEquals(expected.getDisplayName(), actual.getDisplayName());
+        assertEquals(expected.getDescription(), actual.getDescription());
+
+        assertEquals(expected.getKey(), actual.getKey());
+        assertEquals(expected.getValue(), actual.getValue());
+        assertEquals(expected.getLabel(), actual.getLabel());
+    }
+
+    static void assertSecretReferenceConfigurationSettingEquals(SecretReferenceConfigurationSetting expected,
+        SecretReferenceConfigurationSetting actual) {
+        assertEquals(expected.getSecretId(), actual.getSecretId());
+        assertEquals(expected.getKey(), actual.getKey());
+        assertEquals(expected.getValue(), actual.getValue());
+        assertEquals(expected.getLabel(), actual.getLabel());
+    }
+
     /**
      * Helper method to verify that the RestResponse matches what was expected.
      *
@@ -675,5 +918,11 @@ public abstract class ConfigurationClientTestBase extends TestBase {
     static void assertContainsHeaders(HttpHeaders headers, HttpHeaders headerContainer) {
         headers.stream().forEach(httpHeader ->
             assertEquals(headerContainer.getValue(httpHeader.getName()), httpHeader.getValue()));
+    }
+
+    private String getFeatureFlagConfigurationSettingValue(String key) {
+        return "{\"id\":\"" + key + "\",\"description\":null,\"display_name\":\"Feature Flag X\""
+                   + ",\"enabled\":false,\"conditions\":{\"client_filters\":[{\"name\":"
+                   + "\"Microsoft.Percentage\",\"parameters\":{\"Value\":\"30\"}}]}}";
     }
 }